pyproject.toml
[tool.poetry]
name = "air-quality-streaming-labs"
version = "0.1.0"
description = ""
authors = ["Binh Minh <binhm63@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = ">=3.13"
polars = ">=1.36.0,<2.0.0"
pyarrow = ">=22.0.0,<23.0.0"
tqdm = ">=4.67.1,<5.0.0"
plotly = ">=6.5.0,<7.0.0"
numpy = "^2.4.0"
kafka-python = "^2.3.0"
pyspark = "3.5.3"
cassandra-driver = "^3.29.3"
streamlit = "^1.28.0"
pandas = "^2.2.0"
websockets = "^12.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"